requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
pillow>=10.0.0 